First attempt of a buildable CM configuration
This configuration does not work yet.  I am a n00b when it comes to Android,
if you ask me the build system is a mess, would have been so much nicer with
OpenEmbedded :P

As a starting point I looked at this repo:
https://github.com/CM9-Pascal2/android_device_rockchip_rk2918.git

So far I focused on the recovery image, "make recoveryimage" works, but the
resulting image is not yet usable.

Currently it has to be repacked manually with my custom mkbootimg tool,
I did not look into integrating it into the build yet.

If repacked correctly, the rescue image will boot, however the screen
stays black and the rescue application is not working. Adb comes up
though and it is possible to login via adb shell.